compacted MoveCommand into one
