+++ src/main/scala/model/fileIOComponent/fileIO_Json_Impl/FileIO.scala
@@ -36,6 +36,9 @@ class FileIO extends FileIOInterface {
     var col = 0
 
     for (field <- fields.value) {
+      if (row == xPos && col == yPos) {
+        level.dungeon(row)(col) = Field((field \ "fieldvalue").as[Int], (field \ "fieldtype").as[String], (field \ "fog").as[Boolean], true)
+      }
       level.dungeon(row)(col) = Field((field \ "fieldvalue").as[Int], (field \ "fieldtype").as[String], (field \ "fog").as[Boolean], false)
       col += 1
       if (col % size == 0) {
